+++ src/pulumi/core-shared/__main__.py
@@ -3,6 +3,7 @@
 import iam
 import analytics
 import automation
+import kubernetes
 import management
 import network
 import storage