+++ src/pulumi/core-shared/kubernetes/__init__.py
@@ -1,0 +1,6 @@
+# Init the platform outputs
+from project_config import platform_outputs
+
+# Load sub-modules
+from . import cluster
+from . import ssl